+++ Sources/SwiftHTMLClientRuntime/ClientDOMHost.swift
@@ -1,8 +1,8 @@
-public protocol EmbeddedDOMHost {
+public protocol ClientDOMHost {
     associatedtype Node
 
     func createElement(_ tagName: String) -> Node
     func createText(_ text: String) -> Node
-    func setAttribute(_ attribute: EmbeddedHTMLAttribute, on node: Node)
+    func setAttribute(_ attribute: ClientHTMLAttribute, on node: Node)
     func appendChild(_ child: Node, to parent: Node)
 }