+++ Sources/SwiftHTMLClientRuntime/ClientHTMLElement.swift
@@ -1,12 +1,12 @@
-public struct EmbeddedHTMLElement: Sendable, Equatable {
+public struct ClientHTMLElement: Sendable, Equatable {
     public let tagName: String
-    public let attributes: [EmbeddedHTMLAttribute]
-    public let children: [EmbeddedHTMLNode]
+    public let attributes: [ClientHTMLAttribute]
+    public let children: [ClientHTMLNode]
 
     public init(
         _ tagName: String,
-        attributes: [EmbeddedHTMLAttribute] = [],
-        children: [EmbeddedHTMLNode] = []
+        attributes: [ClientHTMLAttribute] = [],
+        children: [ClientHTMLNode] = []
     ) {
         self.tagName = tagName
         self.attributes = attributes
@@ -15,13 +15,13 @@ public struct EmbeddedHTMLElement: Sendable, Equatable {
 
     public init(
         _ tagName: String,
-        _ attributes: EmbeddedHTMLAttribute...,
-        @EmbeddedHTMLBuilder children: () -> [EmbeddedHTMLNode]
+        _ attributes: ClientHTMLAttribute...,
+        @ClientHTMLBuilder children: () -> [ClientHTMLNode]
     ) {
         self.init(tagName, attributes: attributes, children: children())
     }
 
-    public func mount<Host: EmbeddedDOMHost>(
+    public func mount<Host: ClientDOMHost>(
         into host: Host,
         parent: Host.Node
     ) {